pyproject.toml
[project]
name = "web-test-agent"
version = "0.1.0"
description = "AI-powered Web Automation Testing Agent System"
requires-python = ">=3.11,<4.0"
dependencies = [
  "deepagents>=0.5,<0.6",
  "fastapi>=0.115,<1.0",
  "langchain-openai>=1.0,<2.0",
  "langgraph>=1.0,<2.0",
  "python-dotenv>=1.0,<2.0",
  "sse-starlette>=2.0,<3.0",
  "uvicorn[standard]>=0.30,<1.0",
]

[project.scripts]
webtestagent = "webtestagent.cli.main:main"
webtestagent-web = "webtestagent.web.api:serve"
webtestagent-api = "webtestagent.web.api:serve"

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/webtestagent"]

[tool.uv]
package = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "anyio>=4.0",
    "httpx>=0.27",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.10",
]
